rag/requirements.txt
fastapi
uvicorn[standard]
aiokafka
apache-flink
orjson
asyncio
kafka
yfinance
grpcio
grpcio-tools
pandas
pyarrow
python-dotenv
requests
pinecone
google-generativeai
sentence-transformers
google-cloud-storage